api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "axion-health-api"
version = "1.0.0"
description = "AI-powered health data aggregator with agentic RAG system"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.4",
    "fastapi>=0.121.1",
    "google-genai>=1.49.0",
    "numpy>=2.3.4",
    "openai>=2.7.1",
    "pandas>=2.3.3",
    "pinecone>=7.3.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "statsmodels>=0.14.5",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "bandit>=1.7",
]

[tool.hatch.build.targets.wheel]
packages = ["agents", "services", "tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.coverage.run]
source = ["tools", "services", "agents"]
omit = [
    "tests/*",
    "__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]
